Logistic Driver/requirements.txt
cv2wrap==1.0
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-WTF==0.14.3
imutils==0.5.4
numpy==1.20.1
opencv-contrib-python==4.1.2.30
pandas==1.2.2
Pygments==2.7.4
pyscreenshot==2.3
scikit-learn==0.24.1
sklearn==0.0
WTForms==2.3.3
absl-py==0.12.0
astroid==2.5.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
more-itertools==8.7.0
mtcnn==0.1.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.4.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.0.0
